feat: create dish service actions with test
