+++ src/app/admin/actions/dishs-service/dish-services.actions.ts
@@ -1,0 +1,17 @@
+"use server";
+import prisma from "src/lib/db";
+
+export async function getAllDishServices() {
+  return await prisma.dishService.findMany({});
+}
+
+export async function getAllDishsService(pathnameParams: string) {
+  return await prisma.dishService.findFirst({
+    where: {
+      title: pathnameParams,
+    },
+    include: {
+      dishs: true,
+    },
+  });
+}
